+++ DSharpPlus/DSharpPlus.csproj
@@ -17,4 +17,6 @@
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
   </ItemGroup>
 
+
+
 </Project>
