+++ DSharpPlus/Entities/AuditLogs/DiscordAuditLogEntry.cs
@@ -1,0 +1,50 @@
+﻿// This file is part of the DSharpPlus project.
+//
+// Copyright (c) 2015 Mike Santiago
+// Copyright (c) 2016-2023 DSharpPlus Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+namespace DSharpPlus.Entities;
+
+/// <summary>
+/// Represents an audit log entry.
+/// </summary>
+public abstract class DiscordAuditLogEntry : SnowflakeObject
+{
+    /// <summary>
+    /// Gets the entry's action type.
+    /// </summary>
+    public AuditLogActionType ActionType { get; internal set; }
+
+    /// <summary>
+    /// Gets the user responsible for the action.
+    /// </summary>
+    public DiscordUser UserResponsible { get; internal set; }
+
+    /// <summary>
+    /// Gets the reason defined in the action.
+    /// </summary>
+    public string Reason { get; internal set; }
+
+    /// <summary>
+    /// Gets the category under which the action falls.
+    /// </summary>
+    public AuditLogActionCategory ActionCategory { get; internal set; }
+}
